pyproject.toml
[tool.poetry]
name = "django-sssoon"
version = "0.1.9"
description = "A simple Django app to add a beautiful coming soon page to your project."
authors = ["Hareem Adderley <HADDERLEY@KINGPINAPPS.COM>"]

packages = [
    { include = "sssoon" }
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "*"
django-recaptcha = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

